Extract AppBar component to `metabase/nav`
