+++ frontend/src/metabase/App.styled.tsx
@@ -3,8 +3,6 @@ import { css } from "@emotion/react";
 import { color } from "metabase/lib/colors";
 
 import { NAV_HEIGHT } from "metabase/nav/constants";
-import { space } from "metabase/styled-components/theme";
-import Icon from "metabase/components/Icon";
 
 const adminCss = css`
   flex-direction: column;
@@ -22,42 +20,4 @@ export const AppContent = styled.main`
   width: 100%;
   overflow: auto;
   background-color: ${color("content")};
-`;
-
-export const AppBar = styled.div`
-  position: relative;
-  display: flex;
-  align-items: center;
-  width: 100%;
-  background-color: ${color("bg-white")};
-  border-bottom: 1px solid ${color("border")};
-  z-index: 4;
-`;
-
-export const LogoIconWrapper = styled.div`
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 6px;
-  padding: ${space(1)};
-  margin-left: ${space(1)};
-
-  &:hover {
-    background-color: ${color("bg-light")};
-  }
-`;
-
-export const SidebarButton = styled(Icon)`
-  border: 1px solid ${color("border")};
-  padding: ${space(1)};
-  border-radius: 4px;
-  margin-left: ${space(1)};
-  color: ${color("text-medium")};
-
-  &:hover {
-    color: ${color("brand")};
-    border-color: ${color("brand")};
-    cursor: pointer;
-  }
 `;
